requirements.txt
boto3>=1.35.0
elasticsearch>=8.15.0,<10
python-dotenv>=1.0.0
streamlit>=1.39.0
